McsInputStream: terminate if error occurred
This terminates the input stream when an error occurred and does not
wait for the handler thread in the McsService to send the interrupt
signal.

This hopefully fixes a situation that I had where tear down messages
were created in a busy loop because of repeatedly reading -1 from the
input (I don't know how it got into the situation as the log was filled
with the messages from the tear down).