do not hardcode the apk path
The apk path is hardcoded, but it is in fact placed inside the build
path. This patch changes it, so the apk path is calculated from the
build path.

This patch does not change any behavior. However, it allows future
changes to build path to automaticly reflect on the apk path.