ENF: Handle issue with negative or 0 durations due to system time changes
With time synchronisation and can happen that timestamps in database are newer
than reported system time, which then can lead to reported exposure with
multiple measurements but negative or zero reported duration, resulting in
divide-by-zero and other issues in average RSSI calculation. This fixes the
issue, by ignoring new measurements of the same RPI when they are seemingly
older than a previous measurement.