Change translation from "Clique" to "Toque" to make more sense in touch devices
I usually translate Tap to "Clique", which is like click. Mostly because most people don't use "Toque" when talking to someone else and even though it's not a mouse based device, this is usually how it goes. After noticing this (this was a subconscious decision before) I've decided to make translations more consistent with what other translators are doing.

Sorry about the extra commits! I like to proofread my translations a few times and I don't always catch everything, this is relatively minor.