Use cache dir instead of files dir
Installed packages are deleted after installation, but if installation
fails and never retried (for instance: one large package is downloaded
successfully, then a small package is downloaded unsuccessfully due to a
network error, then the user uninstalls the app because it doesn't
work), this can still lead to storage leaks from
downloaded-but-never-installed apps. We should use `cacheDir` instead:
it will allow the system and users to clear the cache themselves.